Auto keyboard input source switcher
This commits contains the seed for an automated keyboard input source switcher.
At the moment it simply prints the source of all HID events from the keybard
device class to the console.

This allows to identify keyboard devices. The next step is to add a
configuration file to be read so a specific device can be associated with a
specific keymap